rename the api to sites
