+++ client/src/components/map/Map/Map.js
@@ -82,7 +82,7 @@ class NCAMap extends Component {
   componentDidMount() {
     this.props.dispatch(
       handleGetSiteData(
-        "/api/location/%7B%22lon%22:-122.643154,%22lat%22:45.55659%7D/500/meters"
+        "/api/location/%7B%22lon%22:-122.553154,%22lat%22:45.55659%7D/1500/meters"
       )
     );
   }
