+++ routes/index.js
@@ -1,4 +1,4 @@
-import {router as api} from "./api";
+import {router as api} from "./sites";
 
 const mountRoutes = (app) => {
   app.use("/api/location", api);
